requirements.txt
streamlit>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
twilio>=8.0.0
